Revert "chore: Refactor folder structure."
This reverts commit 6a88e5ec4215e6d5c70fcae2babf5e112b6a443f.
