--- src/TestData/CONTRIBUTING.md
@@ -1,5 +1,0 @@
-# Contributing Guidelines
-
-## Basic guidelines
-
-Base classes are defined at the top level of the domain while specific implementations grouped in sub-directores.
