+++ src/TestData/Factory.php
@@ -2,7 +2,12 @@
 
 namespace Give\TestData;
 
-abstract class ProviderFactory extends BaseFactory {
+use Faker\Generator;
+
+abstract class Factory {
+
+	/** @var \Faker\Generator */
+	protected $faker;
 
 	protected $providers = [
 		'currency'    => Provider\RandomCurrency::class,
@@ -12,11 +17,22 @@ abstract class ProviderFactory extends BaseFactory {
 	];
 
 	public function __construct() {
-		parent::__construct();
+		$this->faker = \Faker\Factory::create();
 		foreach ( $this->providers as $alias => $providerClass ) {
 			$this->$alias = new $providerClass( $this->faker );
 		}
 	}
+
+	public function make( $count ) {
+		return array_map(
+			function() {
+				return $this->definition();
+			},
+			range( 1, $count )
+		);
+	}
+
+	abstract public function definition();
 
 	/**
 	 * Forward provider calls
