+++ src/TestData/Provider/RandomCurrency.php
@@ -2,8 +2,6 @@
 
 namespace Give\TestData\Provider;
 
-use Give\TestData\RandomProvider;
-
 class RandomCurrency extends RandomProvider {
 
 	/** @var array [currencyCode, ... ] */
