+++ src/TestData/Provider/RandomGateway.php
@@ -2,8 +2,6 @@
 
 namespace Give\TestData\Provider;
 
-use Give\TestData\RandomProvider;
-
 class RandomGateway extends RandomProvider {
 
 	/** @var array [ gatewaySlug, ... ] */
