+++ src/TestData/Provider/RandomPaymentMode.php
@@ -2,8 +2,6 @@
 
 namespace Give\TestData\Provider;
 
-use Give\TestData\RandomProvider;
-
 class RandomPaymentMode extends RandomProvider {
 
 	public function __invoke() {
